test(gui): merge six widget-rule files into one and retire five std::clamp cases
Six files whose combined subject is one sentence -- what a control shows, what
it lets you do, what range it maps, how big its window may be -- and each of
which carried more preamble than propositions. Merged, their include blocks
and namespace scaffolding collapse from six copies to one.

The slider laws become two cases from fifteen. What the split form could not
say is that all three laws owe the same three things: the ends land exactly on
the ends, the map is invertible across its whole range, and it never goes
backwards. Written per law per property, that was fifteen names to read and no
statement of the shared obligation. The anchor table also names each law's
characteristic point, which is the part that is actually different: the log
law's geometric midpoint, and the log-linear law's switch point, where the two
pieces must agree or the slider jumps under the cursor.

Five SliderMappingLinear cases are deleted outright. Their own comment said
they "do NOT call any function from slider_mapping.hpp" -- they asserted that
std::clamp clamps. The convention they stood for is real, but nothing of ours
implements it (SliderWithInput hands the raw value to ImGui), so the coverage
that means anything is the modal-driven gui_test case, which is now named in a
comment where the five cases were.

491 counted lines across six files become 434 in one.
